+++ CHANGELOG.rst
@@ -48,6 +48,9 @@ New Features
 * util-stats: `c.t.finagle.stats.DenylistStatsReceiver` now includes methods for creating
   `DenyListStatsReceiver` from partial functions. ``PHAB_ID=D576833``
 
+* util-core: `c.t.util.FuturePool` now supports exporting the number of its pending tasks via
+  `numPendingTasks`. ``PHAB_ID=D582979``
+
 20.10.0
 -------